Add files via upload
In the Nim Tac Toe game, players each take a piece of the donut and then the first one to win the Nim game gets to place their piece on the tic tac toe board. In the second game along with it called Tictl, players will take turn playing a letter on the board to see who can create a three letter word first. This game uses binary search which was my first time using that algorithm .

In order to play this game on your device, you must have a android phone and developer options enabled on the phone. In order to enable dev options you can watch the following YouTube video: https://youtu.be/gn4bRTFicZw